+++ src/shuuten/_api.py
@@ -5,16 +5,6 @@ from logging import DEBUG, Formatter, Handler, Logger, StreamHandler, getLogger
 from typing import cast
 from uuid import uuid4
 
-from ._destinations import (
-    MSTeamsWebhookDestination,
-    SESDestination,
-    SlackWebhookDestination,
-)
-from ._integrations import (
-    ShuutenContextFilter,
-    ShuutenJSONFormatter,
-    ShuutenNotificationHandler,
-)
 from ._log import LOG, quiet_third_party_logs
 from ._models import Config, Event, NotificationContext, Platform
 from ._notifier import Notifier
@@ -23,6 +13,16 @@ from ._runtime import (
     reset_notification_context,
     reset_runtime_context,
     set_notification_context,
+)
+from .destinations import (
+    MSTeamsWebhookDestination,
+    SESDestination,
+    SlackWebhookDestination,
+)
+from .integrations import (
+    ShuutenContextFilter,
+    ShuutenJSONFormatter,
+    ShuutenNotificationHandler,
 )
 
 _NOTIFIER: Notifier | None = None