+++ src/shuuten/_destinations/__init__.py
@@ -1,9 +1,0 @@
-__all__ = [
-    'SESDestination',
-    'SlackWebhookDestination',
-    'MSTeamsWebhookDestination',
-]
-
-from ._email import SESDestination
-from ._slack import SlackWebhookDestination
-from ._teams import MSTeamsWebhookDestination