+++ src/shuuten/_log.py
@@ -1,4 +1,5 @@
-from logging import WARNING, getLogger
+# noinspection PyProtectedMember
+from logging import ERROR, WARNING, _nameToLevel, getLogger
 
 LOG = getLogger('shuuten')
 
@@ -16,3 +17,10 @@ def quiet_third_party_logs(level: int = WARNING) -> None:
         's3transfer',
     ):
         getLogger(name).setLevel(level)
+
+
+def level_to_int(level: 'str | int') -> int:
+    if isinstance(level, int):
+        return level
+
+    return _nameToLevel.get(level.upper(), ERROR)