+++ src/shuuten/destinations/__init__.py
@@ -1,0 +1,9 @@
+__all__ = [
+    'SESDestination',
+    'SlackWebhookDestination',
+    'MSTeamsWebhookDestination',
+]
+
+from .email import SESDestination
+from .slack import SlackWebhookDestination
+from .teams import MSTeamsWebhookDestination